+++ lib/active_record/connection_adapters/sqlite3_wasm_adapter.rb
@@ -52,6 +52,8 @@ module ActiveRecord
 
         def busy_timeout(...) = nil
 
+        def busy_handler_timeout=(...) = nil
+
         def execute(sql)
           @last_statement = Statement.new(self, sql)
           @last_statement.result